# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payment QR Service API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/terminal/{organization:uuid}": {
            "get": {
                "tags": [
                    "Terminal"
                ],
                "summary": "Вернуть все терминалы по организации",
                "operationId": "b249e6ae953c34d83482927944d94f4f",
                "parameters": [
                    {
                        "name": "organization:uuid",
                        "in": "path",
                        "description": "UUID организации",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {},
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TerminalResource"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Get all child terminal of the user"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Не авторизован."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Общая ошибка сервера.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Error server"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Terminal"
                ],
                "summary": "Создать терминал",
                "operationId": "c996a5fad7f65e3dd0a4e85efee5d37c",
                "parameters": [
                    {
                        "name": "organization:uuid",
                        "in": "path",
                        "description": "UUID организации",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "required": [
                                            "name",
                                            "organization_uuid"
                                        ],
                                        "properties": {
                                            "name": {
                                                "description": "Название терминала, обязательное поле, строка от 3 до 255 символов.",
                                                "type": "string",
                                                "maxLength": 255,
                                                "minLength": 3
                                            },
                                            "organization_uuid": {
                                                "description": "UUID организации, обязательное поле.",
                                                "type": "string",
                                                "format": "uuid"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TerminalResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully create terminal"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка создание терминала.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed create terminal"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Не авторизован."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Общая ошибка сервера.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Error server"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/terminal/{terminal:uuid}": {
            "get": {
                "tags": [
                    "Terminal"
                ],
                "summary": "Вернуть терминал по uuid",
                "operationId": "68c7de6cb20e2e24b155ccedbc730332",
                "parameters": [
                    {
                        "name": "terminal:uuid",
                        "in": "path",
                        "description": "UUID терминала",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {},
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TerminalResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully create terminal"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Не авторизован."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Общая ошибка сервера.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Error server"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TerminalResource": {
                "title": "Terminal Resource",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "uuid": {
                        "type": "string"
                    },
                    "organization_uuid": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}